requirements.txt
fastapi
uvicorn
pydantic
typing
frozendict
mysql-connector-python